Corrections de build

+++ medieval-town/ServiceReceiver.h
@@ -40,7 +40,7 @@ namespace Models
 		void addMarginalService(Service service, Building* building); // Ajoute un bâtiment de service pouvant desservir cet endroit pour un nouvel habitant
 		void removeMarginalService(Service service, Building* building); // Retire le bâtiment de service pouvant desservir cet endroit pour le service donné pour un nouvel habitant
 
-		float getAttractiveness(Pop pop) const; // Retourne l'attractivité pour une population en fonction des services disponibles
+		int getAttractiveness(Pop pop) const; // Retourne l'attractivité pour une population en fonction des services disponibles
 
 	protected:
 		virtual void updateAttractiveness(); // Met à jour l'attractivité après que les services disponibles aient évolués (virtual pour laisser Tile l'override)
@@ -50,7 +50,7 @@ namespace Models
 
 		std::map<Service, std::vector<Building*>> marginalServiceBuildings; // Pour chaque service, liste des bâtiments de service à portée pouvant accueillir un habitant de plus, afin de calculer l'attractivité. Privée car toute modification doit passer par les setters afin que soit appelé updateAttractiveness()
 
-		std::map<Pop, float> attractiveness; // Attractivité pour une population en fonction des services disponibles
+		std::map<Pop, int> attractiveness; // Attractivité pour une population en fonction des services disponibles
 	};
 }
